+++ src/test/scala/org/clapper/markwrap/BaseSpec.scala
@@ -1,0 +1,5 @@
+package org.clapper.markwrap
+
+import org.scalatest.{FlatSpec, Matchers}
+
+trait BaseSpec extends FlatSpec with Matchers